214 Scheduled job now knows how to refresh token as well
Last token used for call auth stored and accepted until next call
